[kyo-system] name the stream sinks after the module and expose the write options
Rename StreamFileExtensions to StreamSystemExtensions, matching the
StreamCoreExtensions naming in kyo-core, and move the sink tests out of
PathTest into StreamSystemExtensionsTest.

writeTo and writeLinesTo now take append and createFolders, with the same
names and defaults as the Path write methods. The failure cleanup removes
the file only when append is false, since an appending sink would otherwise
delete content it never wrote.

Scala permits default arguments on a single alternative of an overloaded
method, so the Stream[String, S] sink keeps the defaults and the
Stream[Byte, S] sink states the options as overloads.
