+++ kyo-system/shared/src/test/scala/kyo/PathTest.scala
@@ -1410,72 +1410,6 @@ class PathTest extends kyo.test.Test[Any]:
     }
 
     // =========================================================================
-    // Stream extensions
-    // =========================================================================
-
-    "Stream[Byte].writeTo creates file with correct byte content" in {
-        val bytes = Array[Byte](10, 20, 30, 40, 50)
-        for
-            dir <- Path.tempDir("kyo-path-stream-test")
-            file = dir / "stream-byte-write.bin"
-            _      <- Scope.run(Stream.init(Chunk.from(bytes)).writeTo(file))
-            result <- file.readBytes
-            _      <- dir.removeAll
-        yield assert(result.toArray.toList == bytes.toList)
-        end for
-    }
-
-    "Stream[String].writeTo writes concatenated strings" in {
-        val parts = List("hello", ", ", "world")
-        for
-            dir <- Path.tempDir("kyo-path-stream-test")
-            file = dir / "stream-string-write.txt"
-            _      <- Scope.run(Stream.init(Chunk.from(parts)).writeTo(file))
-            result <- file.read
-            _      <- dir.removeAll
-        yield assert(result == "hello, world")
-        end for
-    }
-
-    "Stream[String].writeLinesTo writes each element as a line" in {
-        val lines = Chunk("alpha", "beta", "gamma")
-        for
-            dir <- Path.tempDir("kyo-path-stream-test")
-            file = dir / "stream-lines-write.txt"
-            _      <- Scope.run(Stream.init(lines).writeLinesTo(file))
-            result <- file.readLines
-            _      <- dir.removeAll
-        yield assert(result == lines)
-        end for
-    }
-
-    "Stream[String].writeTo with ISO-8859-1 charset encodes correctly" in {
-        val charset = StandardCharsets.ISO_8859_1
-        val text    = "caf\u00e9"
-        for
-            dir <- Path.tempDir("kyo-path-stream-test")
-            file = dir / "stream-charset-write.txt"
-            _      <- Scope.run(Stream.init(Chunk(text)).writeTo(file, charset))
-            result <- file.read(charset)
-            _      <- dir.removeAll
-        yield assert(result == text)
-        end for
-    }
-
-    "Stream[String].writeLinesTo with ISO-8859-1 charset encodes correctly" in {
-        val charset = StandardCharsets.ISO_8859_1
-        val lines   = Chunk("pr\u00e9", "deux\u00e8me")
-        for
-            dir <- Path.tempDir("kyo-path-stream-test")
-            file = dir / "stream-lines-charset-write.txt"
-            _      <- Scope.run(Stream.init(lines).writeLinesTo(file, charset))
-            result <- file.readLines(charset)
-            _      <- dir.removeAll
-        yield assert(result == lines)
-        end for
-    }
-
-    // =========================================================================
     // Edge cases
     // =========================================================================
 
@@ -2001,30 +1935,6 @@ class PathTest extends kyo.test.Test[Any]:
         }
     }
 
-    "writeTo with empty byte stream creates an empty file" in {
-        for
-            dir <- Path.tempDir("kyo-test")
-            file = dir / "empty-byte-stream.bin"
-            _      <- Scope.run(Stream.empty[Byte].writeTo(file))
-            exists <- file.exists
-            bytes  <- file.readBytes
-            _      <- dir.removeAll
-        yield assert(exists && bytes.isEmpty)
-        end for
-    }
-
-    "writeLinesTo with empty stream creates an empty file" in {
-        for
-            dir <- Path.tempDir("kyo-test")
-            file = dir / "empty-lines-stream.txt"
-            _      <- Scope.run(Stream.empty[String].writeLinesTo(file))
-            exists <- file.exists
-            bytes  <- file.readBytes
-            _      <- dir.removeAll
-        yield assert(exists && bytes.isEmpty)
-        end for
-    }
-
     // =========================================================================
     // name / extName edge cases
     // =========================================================================
@@ -2055,37 +1965,6 @@ class PathTest extends kyo.test.Test[Any]:
     // =========================================================================
     // Regression tests — inspired by known issues in fs2, os-lib, and zio-process
     // =========================================================================
-
-    // Inspired by fs2 #3667: writeTo should not leave a file containing partial data
-    // when the input stream fails mid-flight.
-    "writeTo does not leave file with partial data when stream fails mid-flight" in {
-        for
-            dir <- Path.tempDir("kyo-test")
-            file = dir / "should-not-have-partial.txt"
-            result <- Abort.run[FileWriteException] {
-                Scope.run {
-                    // A stream that emits one chunk then fails with a FileIOException
-                    val badStream: Stream[Byte, Abort[FileWriteException]] =
-                        Stream.init(Chunk[Byte](1, 2, 3)).concat(
-                            Stream[Byte, Abort[FileWriteException]](
-                                Abort.fail(FileIOException(file, new java.io.IOException("stream error")))
-                            )
-                        )
-                    badStream.writeTo(file)
-                }
-            }
-            exists <- file.exists
-            bytes  <- Abort.run[FileReadException](file.readBytes)
-            _      <- dir.removeAll
-        yield
-            assert(result.isFailure)
-            // The file should either not exist or be empty — not contain partial data.
-            // If this assertion fails it means partial data was written and left behind.
-            bytes match
-                case Result.Success(b) => assert(b.isEmpty, s"Partial data left in file: ${b.size} bytes")
-                case Result.Failure(_) => () // file doesn't exist — also acceptable
-        end for
-    }
 
     // Inspired by fs2 #1005: buffer reuse corruption after rechunking.
     // Each chunk read from readBytesStream must be an independent copy, not aliased
@@ -2172,32 +2051,6 @@ class PathTest extends kyo.test.Test[Any]:
         yield
             assert(firstChunk.nonEmpty)
             assert(finalContent == "overwritten")
-        end for
-    }
-
-    // Inspired by fs2 #3667: writeTo with a failing mid-stream should not leave corrupt partial data.
-    "writeTo with failing stream does not leave corrupt partial file" in {
-        for
-            dir <- Path.tempDir("kyo-writeto-fail")
-            file = dir / "partial.bin"
-            result <- Abort.run[FileWriteException] {
-                Scope.run {
-                    val failingStream: Stream[Byte, Abort[FileWriteException]] =
-                        Stream.init(Chunk[Byte](10, 20, 30)).concat(
-                            Stream[Byte, Abort[FileWriteException]](
-                                Abort.fail(FileIOException(file, new java.io.IOException("mid-stream error")))
-                            )
-                        )
-                    failingStream.writeTo(file)
-                }
-            }
-            bytes <- Abort.run[FileReadException](file.readBytes)
-            _     <- dir.removeAll
-        yield
-            assert(result.isFailure)
-            bytes match
-                case Result.Success(b) => assert(b.isEmpty, s"Partial data found: ${b.size} bytes")
-                case Result.Failure(_) => ()
         end for
     }
